Refactor folder structure for clarity and simplicity
This commit refactors the project's folder structure to improve clarity, reduce redundancy, and align with the architecture described in the README.

Key changes include:
- Merged the `src/alerts` and `src/email_alerts` directories into a single `src/alerts` module.
- Consolidated the data provider implementations into a `src/data` directory, removing the simulated `src/data_provider` and merging duplicate `FyersDataProvider` implementations.
- Renamed `src/order` to `src/execution` to better reflect its purpose and match the documentation.
- Removed the empty `src/api` directory.
