--- src/api/__init__.py
@@ -1,3 +1,0 @@
-"""
-API Package for Trading System
-"""
