--- src/data_provider/__init__.py
@@ -1,3 +1,0 @@
-"""
-Data Provider Module
-"""
