--- src/email_alerts/__init__.py
@@ -1,3 +1,0 @@
-"""
-Email Alerts Module
-"""
