+++ src/execution/__init__.py
@@ -1,1 +1,3 @@
-# Trading execution module
+"""
+Execution Module
+"""
