--- src/order/__init__.py
@@ -1,3 +1,0 @@
-"""
-Order Management Module
-"""
