+++ tests/test_user_aware_simple.py
@@ -8,7 +8,7 @@ from datetime import datetime
 from flask_bcrypt import Bcrypt
 
 # Add src to path for imports
-sys.path.insert(0, os.path.join(os.path.dirname(__file__), '..', 'src'))
+sys.path.insert(0, os.path.abspath(os.path.join(os.path.dirname(__file__), '..', 'src')))
 
 from datastore.database import get_database_manager
 from datastore.models import User, Order, Trade, Position, Strategy, SuggestedStock, Configuration, Log
